Apply PR review round 1 findings
Brand scoping is now opt-in on the product-labels factory: supplements keep
it, foods always use the generic path (no 2M-row brand preload, no exclusive
scope hiding generic rows). Batch search help no longer claims FDC id/UPC
lookup (single search-labels handles those). Route logic shared via one
server-only product-labels route helper; both API routes are thin bindings.
